docker_pyt/requirements.txt
numpy==2.1.0
chromadb
google-generativeai
pypdf